// src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  CHAIRMAN = 'CHAIRMAN',
  SECRETARY = 'SECRETARY',
  TREASURER = 'TREASURER',
  JOINT_SECRETARY = 'JOINT_SECRETARY',
  COMMITTEE_MEMBER = 'COMMITTEE_MEMBER',
  TOWER_CAPTAIN = 'TOWER_CAPTAIN',
  ACCOUNTANT = 'ACCOUNTANT',
  FACILITY_MANAGER = 'FACILITY_MANAGER',
  SECURITY_SUPERVISOR = 'SECURITY_SUPERVISOR',
  SECURITY_GUARD = 'SECURITY_GUARD',
  MAINTENANCE_STAFF = 'MAINTENANCE_STAFF',
  OWNER_RESIDENT = 'OWNER_RESIDENT',
  OWNER_NONRESIDENT = 'OWNER_NONRESIDENT',
  TENANT = 'TENANT',
  FAMILY_MEMBER = 'FAMILY_MEMBER',
}